requirements.txt
# ReadingTypeID Agent 依赖包

# 核心依赖
openai>=1.10.0
pandas>=2.0.0
python-dotenv>=1.0.0

# XML处理
lxml>=4.9.0

# 可选：增强功能
requests>=2.28.0
urllib3>=1.26.0

# 开发依赖 (可选)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0 